Move test helpers only used for testing fake marathon to tests
